[SW-4944] tests: Add infrastructure for running on multiple devices
Add an option of running a tests group on multiple devices, by creating
a thread for each device.
The threads get synced after a test setup function and before the actual
test starts.

The usage of a single threads-dispatcher function, in hlthunk_tests.c,
forces the usage of the internal _cmocka_run_group_tests() function,
instead of the cmocka_run_group_tests() function, because some
preprocessor operations (e.g. stringizing group name) are not possible.
As this change lead to more code duplication in the various main()
functions, new macros are added to wrap these functions and their
surroundings.

[SW-5053] tests: do not set CMOCKA_TEST_ABORT to 1

It causes seg fault in tests that fail instead of showing the FAILED in
the output.

Change-Id: I98c06f469650c474569e39393fc2d00094d446ea
Signed-off-by: Oded Gabbay <oded.gabbay@gmail.com>