[SW-6870] hlthunk_err_inject: implement the error inject thermal routines
This commit implements the error injection temperature library
routine.
It does so by modifying the temperature offset of the first
temperature sensor found in the device under test and provides
a routine to restore the original offset value as long as no
modification to the value of max-temperature were made in between.

Change-Id: I83da80fc13ad2724e3b2ef340cfaf963b83f91e1
Signed-off-by: Moti Haimovski <mhaimovski@habana.ai>