[SW-13710] tests/gaudi: Add MME convolution test
Add basic MME convolution test to hlthunk,
which program the internal engines registers of MME
and not just send jobs to QM as done in other tests.
This test will make MME engines execute a basic convolution
on specific inputs and weights, then it'll compare the result
with the expected one.

Change-Id: I096bd3f832faa6b004709d310c9c10e37fe343c9
Signed-off-by: farah kassabri <fkassabri@habana.ai>