[SW-0] tests: Open debugfs for the acquired device in root tests
If no PCI address is provided, the setup function of root tests will use
0 as the device index when opening the debugfs nodes, even if a FD of
another device was opened.
The patch fixes it and moves the responsibility of selecting the device
index from debugfs_open() to its callers.

As hltests_root_debug_setup() can now fail after a successful
allocation of the tests state and before calling debugfs_open(), it is
required to initialize the debugfs FDs before debugfs_close() is called
as part of the tear-down function.

Change-Id: I88e5e9ddd6d74076250e0c172d708ca69a2464b6
Signed-off-by: Tomer Tayar <ttayar@habana.ai>