[SW-25736] hlthunk: define interface to get FD for DMA-BUF
We add a new feature to map a dma-buf object to a memory allocation on
the device memory.

The device memory is allocated using our existing allocation API,
where the user will get a handle that represents the allocation.

The user will then need to call the new
API (hlthunk_device_memory_export_dmabuf_fd) and give the handle as
a parameter.

The driver will return a FD that represents the DMA-BUF object that
was created to match that allocation.

Note that in Habana's ASICs, the device memory is *always* pinned,
so there is no requirement for the user (importer) to have ODP
capabilities.

Also note that in GAUDI we don't have MMU capabilities in the device
memory and the user works on physical addresses. Therefore, the user
doesn't pass through the kernel driver to allocate memory there. As
a result, only in GAUDI this new API expects a device memory physical
address (instead of a handle) and a size.

Change-Id: Id808971b4160b08567a3cb7aa4f00163632af185
Signed-off-by: Oded Gabbay <ogabbay@habana.ai>