[SW-17257] tests: implement encapsulated signals tests
Add single and multi-threaded tests for basic testing
of the feature.
Add one test which cause sob wraparound during the
reservation, and eventually cause the reservation
ioctl to fail. then send stagged cs to with encaps
signaling to both sobs, then try to reserve
again and see if reservation passes.

Change-Id: Ibd020f1e5ee52e8b3d48225acf86306083036cf5
Signed-off-by: farah kassabri <fkassabri@habana.ai>