[SW-42330] hl-thunk: ensure errno is not lost during open
During hlthunk_open, when opening using device name, the lib will
attempt to open multiple driver minors, until at least one is
successful and matches device_name.

Hence, if the open fails, for any reason, it will continue to try
and open other minors, that may not exist. In this case, the original
failure reason (errno) will be overwritten by the ENOENT failures
from attempts to access the non-existent minors. This means, the
caller will not be able to see why it failed, instead it will always
get ENOENT errno.

Change-Id: I7aa1a3dc9b48cc1484edf76414fb244097b71034
Signed-off-by: Yuri Nudelman <ynudelman@habana.ai>