[SW-50289] gaudi: temporary update of test_strided_dma
There's a bug that causes test_strided_dma test to fail.
Due to the difficulty of reproducing the bug, we must change
the test such that it will print excess debug info upon failure.
As well, we want to try and run the test again - immediately
after its initial failure.
In such a case, we expect it to similarly fail again, which might
indicate of a non responding DMA engine.

Change-Id: Icf7082a78ba403b6df96f47219f91a651c9af413
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>