[SW-34090] tests: add get_first_avail_interrupt function
In a later additional commit, some user interrupts are reserved for
working with the scheduler ARC, so tests can't use directly the
first_available_interrupt_id field of hw_ip_info.
Instead, use a dedicated function for it, and in this later additional
commit, these reserved interrupts will be taken into consideration
inside the function.

Change-Id: Ief17c1514806e79e10be7770eff9eaea9970051a
Signed-off-by: Tomer Tayar <ttayar@habana.ai>