[SW-48030] add "in device creation" status
On init, the disabled state is cleared right before hw_init and that
causes the device to report on "Operational" state before the nic_init.
This causes errors in some NIC scripts because the NIC ports status is
checked before if the device is operational meaning before the NIC is
operational.
In order to avoid this, a new state "in device creation" is indroduced to
be reported when the device is not disabled but is still in init flow.

Change-Id: I0df3e423a1468bd1580a9f9c36b5b5a70602ac82
Signed-off-by: Omer Shpigelman <oshpigelman@habana.ai>